+++ attributes/repos.rb
@@ -17,7 +17,7 @@
 # limitations under the License.
 #
 
-node.default['system_core']['repos']['ol7_base_latest']['url']     = 'http://yum$ociregion.oracle.com/repo/OracleLinux/OL7/latest/$basearch/'
-node.default['system_core']['repos']['ol7_uek4_latest']['url']     = 'http://yum$ociregion.oracle.com/repo/OracleLinux/OL7/UEKR4/$basearch/'
-node.default['system_core']['repos']['ol7_uek5_latest']['url']     = 'http://yum$ociregion.oracle.com/repo/OracleLinux/OL7/UEKR5/$basearch/'
-node.default['system_core']['repos']['ol7_optional_latest']['url'] = 'http://yum$ociregion.oracle.com/repo/OracleLinux/OL7/optional/latest/$basearch/'
+node.default['system_core']['repos']['ol7_base_latest']['url']     = 'http://yum.oracle.com/repo/OracleLinux/OL7/latest/$basearch/'
+node.default['system_core']['repos']['ol7_uek4_latest']['url']     = 'http://yum.oracle.com/repo/OracleLinux/OL7/UEKR4/$basearch/'
+node.default['system_core']['repos']['ol7_uek5_latest']['url']     = 'http://yum.oracle.com/repo/OracleLinux/OL7/UEKR5/$basearch/'
+node.default['system_core']['repos']['ol7_optional_latest']['url'] = 'http://yum.oracle.com/repo/OracleLinux/OL7/optional/latest/$basearch/'