+++ files/default/install-aws-cli.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/bin/bash -e
 
 # TODO: Fix this error:
 # /tmp/install-aws-cli.sh: line 3: [: -e: binary operator expected
@@ -8,10 +8,24 @@
 # https://urllib3.readthedocs.org/en/latest/security.html#insecureplatformwarning.
 # InsecurePlatformWarning
 
-if [ ! -e ./get-pip.py ] ; then
-  # Only download if it hasn't already been downloaded
-  curl "https://bootstrap.pypa.io/get-pip.py" -o "get-pip.py"
-  python ./get-pip.py
+PIPCMD=''
+if which pip3 >/dev/null ; then
+  PIPCMD="$(which pip3)"
+elif which pip >/dev/null ; then
+  PIPCMD="$(which pip)"
+fi
+
+PYCMD=python
+if which python3 >/dev/null ; then
+  PYCMD=python3
+fi
+
+if [ -z $PIPCMD ] ; then
+  if [ ! -e ./get-pip.py ] ; then
+    # Only download if it hasn't already been downloaded
+    curl "https://bootstrap.pypa.io/get-pip.py" -o "get-pip.py"
+  fi
+  $PYCMD ./get-pip.py
 fi
 
 # Fix for Ubuntu which puts pip into /usr/local/bin but which doesn't have /usr/local/bin
@@ -19,4 +33,4 @@ fi
 if ! env | grep ^PATH | grep '/usr/local/bin' > /dev/null ; then
   export PATH=$PATH:/usr/local/bin
 fi
-pip install awscli
+$PIPCMD install awscli